factor ReturningTrait into ReturningInterface
