+++ src/Common/ReturningInterface.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Aura\Sql_Query\Common;
+
+interface ReturningInterface
+{
+    /**
+     *
+     * Adds returning columns to the query.
+     *
+     * Multiple calls to returning() will append to the list of columns, not
+     * overwrite the previous columns.
+     *
+     * @param array $cols The column(s) to add to the query.
+     *
+     * @return $this
+     *
+     */
+    public function returning(array $cols);
+}
